test(auth): split dashboard/gateway/daemon command tests (#1391)
